+++ src/components/llm_list.py
@@ -40,6 +40,11 @@ class LLMList(tk.Frame):
         cwd = os.getcwd() # current directory
         parent = os.path.dirname(cwd)# parent directory
         model_dir = os.path.join(parent, "model_files")
+
+        # Create model_files directory if it doesn't exist
+        if not os.path.exists(model_dir):
+            os.makedirs(model_dir)
+            return folder_list 
         # check if model_files
         for folder in os.listdir(model_dir): # list all folders in model_files folder
             if os.path.isdir(os.path.join(model_dir, folder)):
@@ -48,9 +53,14 @@ class LLMList(tk.Frame):
 
     def write_list(self, list):
 <<<<<<< HEAD
+<<<<<<< HEAD
+=======
+        self.list.delete(0,tk.END) # delete existing entries
+>>>>>>> 18b5a80842d2d547646a5a0f221bdf64cdc5890d
         count = 0
-        for i in list:
+        for model in list:
             count += 1
+<<<<<<< HEAD
             self.list.insert(count, i)
 =======
         self.list.delete(0,tk.END) # delete existing entries
@@ -67,3 +77,6 @@ class LLMList(tk.Frame):
 
 
 >>>>>>> a0610ba1835c74dcf76d954652a1a43adb5e15a1
+=======
+            self.list.insert(count, model)
+>>>>>>> 18b5a80842d2d547646a5a0f221bdf64cdc5890d